+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - file URl are now properly handled, they must start with `file:///` instead of `file://`
 - Updater do not rely on yaml anymore. Instead, a basic configuration for logging is used. This should prevent random yaml upgrade issues.
 - Do not request VBA caller in threading context (as it cannot be retrieved and result in COM call failure).
+- The `PyxelRestAddIn` does not read default `xlwings` configuration anymore. Allowing to run in addition of the `xlwings` add-in.
 
 ### Removed
 - Drop support for `python` < `3.8`.